+++ platform/camera.h
@@ -21,6 +21,7 @@ typedef struct image image_t;
 typedef enum {
     ESP_VISION_CAMERA_FRAMESIZE_QQVGA = 0,
     ESP_VISION_CAMERA_FRAMESIZE_QVGA,
+    ESP_VISION_CAMERA_FRAMESIZE_VGA,
 } esp_vision_camera_framesize_t;
 
 typedef struct {